requirements.txt
Flask==3.0.3
Flask_RESTful==0.3.10
PyYAML==6.0.1
Requests==2.32.3
